Move compat renderer code to compat sub package
