+++ compat/LICENSE
@@ -1,6 +1,6 @@
 This directory includes source copied and inspired by other packages. Files that
 use code and APIs from other projects are called out at the top of the
-individual file. There licenses are provided here.
+individual file. Their licenses are provided here.
 
 ---
