+++ compat/scripts/mangle.json
