+++ compat/test/compat-shallow-renderer/PreactShallowRendererExtra_test.tsx
