+++ compat/test/compat-shallow-renderer/PreactShallowRendererHooks_test.tsx
