+++ compat/test/compat-shallow-renderer/PreactShallowRendererMemo_test.tsx
