+++ compat/test/compat-shallow-renderer/PreactShallowRenderer_test.tsx
