--- compat/test/index_test.ts
@@ -1,8 +1,0 @@
-import { assert } from 'chai';
-import { HelloWorld } from '../src/index.js';
-
-describe('HelloWorld', () => {
-  it('exports hello world', () => {
-    assert.equal(HelloWorld, 1);
-  });
-});
