+++ package.json
@@ -44,9 +44,9 @@
   },
   "scripts": {
     "build": "tsc --build tsconfig.json",
-    "postbuild": "node ./scripts/transform-internal-properties.mjs",
+    "postbuild": "node ./compat/scripts/transform-internal-properties.mjs",
     "build-cjs": "tsc --build tsconfig-cjs.json",
-    "postbuild-cjs": "node ./scripts/transform-internal-properties.mjs",
+    "postbuild-cjs": "node ./compat/scripts/transform-internal-properties.mjs",
     "clean": "rm -rf build build-cjs",
     "checkformatting": "prettier --check \"**/*.{js,jsx,ts,tsx,mjs,cjs,yml}\"",
     "format": "prettier --list-different --write \"**/*.{js,jsx,ts,tsx,mjs,cjs,yml}\"",
@@ -55,10 +55,10 @@
     "lint-staged": "lint-staged",
     "prepublish": "rm -rf build && yarn build && yarn build-cjs",
     "pretest": "yarn build && echo '{\"type\":\"module\"}' > build/package.json",
-    "test": "nyc mocha -r build/test/init.js build/test/*.js && nyc mocha build/test/compat-shallow-renderer && nyc mocha build/compat/test",
+    "test": "nyc mocha -r build/test/init.js build/test/*.js && nyc mocha --recursive build/compat/test",
     "posttest": "rm build/package.json",
     "pretest-cjs": "yarn build-cjs",
-    "test-cjs": "nyc mocha -r build-cjs/test/init.js build-cjs/test/*.js && nyc mocha build-cjs/test/compat-shallow-renderer && nyc mocha build-cjs/compat/test",
+    "test-cjs": "nyc mocha -r build-cjs/test/init.js build-cjs/test/*.js && nyc mocha --recursive build-cjs/compat/test",
     "test:compat": "yarn test --preact-compat-lib preact/compat",
     "prepare": "husky install"
   },
